Move exponential filter definition to inline header
